pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygemstone"
version = "0.0.1"
description = "Python library for Gemstone Lights permanent-Christmas-lights controllers (AWS-backed cloud API)"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Stefan Slivinski", email = "sslivins@gmail.com" }
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Home Automation"
]
# Runtime dependencies
#
# pycognito provides the AWS Cognito User-Pool SRP auth flow without
# pulling in the full boto3 stack. It is synchronous; we wrap it via
# asyncio.to_thread for the async client surface.
dependencies = [
    "aiohttp>=3.11.12",
    "pycognito>=2024.5.1"
]

[project.optional-dependencies]
tests = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "aioresponses>=0.7.8",
    "python-dotenv>=1.0.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0"
]

[project.urls]
"Homepage" = "https://github.com/sslivins/pygemstone"
"Bug Tracker" = "https://github.com/sslivins/pygemstone/issues"

[project.scripts]
pygemstone = "pygemstone.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.11"
strict_optional = true
warn_unused_ignores = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["pycognito", "pycognito.*"]
ignore_missing_imports = true
